requirements.txt
python-docx
python-pptx
PyMuPDF
xlwings
pywin32
Pillow